Write test queue to file in reverse order
Tests are pulled from the end of the file first. So when we enqueue
them, we should put the first test at the end of the file.

This is important when using the `FINE_GRAIN_TEST_RUNTIME_LOGGER` option
to run slowest tests first. 2ad6ffde75fe8ea8b052d44c1084153433ffc906
actually made it so that slowest tests run last by mistake.